# lib/api-zod/src/generated/types/fileUploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface FileUploadResponse {
  url: string;
  fileName: string;
  fileSize: number;
  mimeType: string;
}